+++ src/icalendar/tests/test_fixed_issues.py
@@ -478,3 +478,15 @@ END:VCALENDAR"""
             expected_tzname = 'Brasília standard'.encode('ascii', 'replace')
         self.assertEqual(dtstart.tzinfo.zone, expected_zone)
         self.assertEqual(dtstart.tzname(), expected_tzname)
+
+    def test_issue_345(self):
+        """Issue #345 - Why is tools.UIDGenerator a class (that must be instantiated) instead of a module? """
+        uid1 = icalendar.tools.UIDGenerator.uid()
+        uid2 = icalendar.tools.UIDGenerator.uid('test.test')
+        uid3 = icalendar.tools.UIDGenerator.uid(unique='123')
+        uid4 = icalendar.tools.UIDGenerator.uid('test.test', '123')
+
+        self.assertEqual(uid1.split('@')[1], 'example.com')
+        self.assertEqual(uid2.split('@')[1], 'test.test')
+        self.assertEqual(uid3.split('-')[1], '123@example.com')
+        self.assertEqual(uid4.split('-')[1], '123@test.test')