+++ src/icalendar/tools.py
@@ -15,19 +15,21 @@ class UIDGenerator(object):
     """
     chars = list(ascii_letters + digits)
 
-    def rnd_string(self, length=16):
+    @staticmethod
+    def rnd_string(length=16):
         """Generates a string with random characters of length.
         """
-        return ''.join([random.choice(self.chars) for _ in range(length)])
+        return ''.join([random.choice(UIDGenerator.chars) for _ in range(length)])
 
-    def uid(self, host_name='example.com', unique=''):
+    @staticmethod
+    def uid(host_name='example.com', unique=''):
         """Generates a unique id consisting of:
             datetime-uniquevalue@host.
         Like:
             20050105T225746Z-HKtJMqUgdO0jDUwm@example.com
         """
         host_name = to_unicode(host_name)
-        unique = unique or self.rnd_string()
+        unique = unique or UIDGenerator.rnd_string()
         today = to_unicode(vDatetime(datetime.today()).to_ical())
         return vText('%s-%s@%s' % (today,
                                    unique,