+++ CHANGES.rst
@@ -5,7 +5,11 @@ Changelog
 5.0.0a2 (unreleased)
 --------------------
 
-- Nothing changed yet.
+Bug fixes:
+
+- Changed tools.UIDGenerator instance methods to static methods
+  Ref: #345
+  [spralja]
 
 
 5.0.0a1 (2022-07-11)